agregacion cEspPoint y qualifyCEspecifico para calificar los conocimientos especificos del aspirante. Cambio a DOUBLE de puntuacion del aspirante. Inconsistencia con sql y DAOAspirante.

+++ SISCONDOC/src/proceso/Aspirante.java
@@ -20,7 +20,7 @@ package proceso;
     private String name;
     private String lastname;
     private String city;
-    private int puntaje;
+    private double puntaje;
     private Convocatoria call;
     private String genero;
     private String jornada;
@@ -110,14 +110,14 @@ package proceso;
     /**
      * @return the puntaje
      */
-    public int getPuntaje() {
+    public double getPuntaje() {
         return puntaje;
     }
 
     /**
      * @param puntaje puntaje set
      */
-    public void setPuntaje(int puntaje) {
+    public void setPuntaje(double puntaje) {
         this.puntaje= puntaje;
     }
